refactor: Update folder structure for better organization template
