+++ app/Enums/TransactionType.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Enums;
 
 enum TransactionType: string
 {
